flatten header array so it's easier to access headers
