+++ event/http.go
@@ -2,7 +2,7 @@ package event
 
 // HTTPEvent is a event schema used for sending events to HTTP subscriptions.
 type HTTPEvent struct {
-	Headers map[string][]string `json:"headers"`
+	Headers map[string]string   `json:"headers"`
 	Query   map[string][]string `json:"query"`
 	Body    interface{}         `json:"body"`
 	Host    string              `json:"host"`
